+++ samples/TurtlePath.Samples.Basic/Application/Handlers/CreateCustomerCommandHandler.cs
@@ -5,7 +5,7 @@ using TurtlePath.Samples.Basic.Domain.Entities;
 
 namespace TurtlePath.Samples.Basic.Application.Handlers;
 
-public sealed class CreateCustomerCommandHandler : BaseEntityCreateCommandHandler<CreateCustomerRequest, CustomerResponse, Customer>
+public sealed class CreateCustomerCommandHandler : CreateCommandHandler<CreateCustomerRequest, CustomerResponse, Customer>
 {
     public CreateCustomerCommandHandler(IServiceProvider serviceProvider)
         : base(serviceProvider)