+++ src/TurtlePath/Commands/DeleteCommandHandler.cs
@@ -4,7 +4,6 @@ namespace TurtlePath.Commands
     using TurtlePath.Exceptions;
     using TurtlePath.Hooks;
     using TurtlePath.Models.Requests;
-    using TurtlePath.Models.Responses;
     using TurtlePath.Persistence;
     using TurtlePath.Validation;
     using TurtlePath.Mapping;
@@ -13,21 +12,21 @@ namespace TurtlePath.Commands
     using Pelican.Mediator;
 
     /// <summary>
-    /// Provides a base implementation for handling patch commands for TurtlePath BaseEntity instances with CId identifiers.
+    /// Provides a base implementation for handling delete commands for TurtlePath BaseEntity instances with CId identifiers.
     /// </summary>
     /// <typeparam name="TRequest">The type of the request.</typeparam>
-    /// <typeparam name="TEntity">The type of the entity being patched.</typeparam>
+    /// <typeparam name="TEntity">The type of the entity being deleted.</typeparam>
     /// <typeparam name="TResponse">The type of the response.</typeparam>
-    public abstract class BaseEntityPatchCommandHandler<TRequest, TResponse, TEntity> : EntityPatchCommandHandler<TRequest, TResponse, TEntity, CId>
+    public abstract class DeleteCommandHandler<TRequest, TResponse, TEntity> : GenericDeleteCommandHandler<TRequest, TResponse, TEntity, CId>
         where TRequest : BaseRequest, IRequest<TResponse>
+        where TResponse : class
         where TEntity : BaseEntity
-        where TResponse : BaseResponse
     {
         /// <summary>
         /// Initializes a new instance of this class.
         /// </summary>
         /// <param name="serviceProvider">The service provider used to resolve dependencies.</param>
-        protected BaseEntityPatchCommandHandler(IServiceProvider serviceProvider) : base(serviceProvider)
+        protected DeleteCommandHandler(IServiceProvider serviceProvider) : base(serviceProvider)
         {
         }
     }