+++ src/TurtlePath/Commands/PatchCommandHandler.cs
@@ -1,7 +1,9 @@
 namespace TurtlePath.Commands
 {
     using Microsoft.Extensions.DependencyInjection;
+    using TurtlePath.Exceptions;
     using TurtlePath.Hooks;
+    using TurtlePath.Models.Requests;
     using TurtlePath.Models.Responses;
     using TurtlePath.Persistence;
     using TurtlePath.Validation;
@@ -9,16 +11,15 @@ namespace TurtlePath.Commands
     using TurtlePath.Domain.Contracts;
     using TurtlePath.Domain.Identifier;
     using Pelican.Mediator;
-    using System;
 
     /// <summary>
-    /// Provides a base implementation for handling create commands for TurtlePath BaseEntity instances with CId identifiers.
+    /// Provides a base implementation for handling patch commands for TurtlePath BaseEntity instances with CId identifiers.
     /// </summary>
     /// <typeparam name="TRequest">The type of the request.</typeparam>
-    /// <typeparam name="TEntity">The type of the entity being created.</typeparam>
+    /// <typeparam name="TEntity">The type of the entity being patched.</typeparam>
     /// <typeparam name="TResponse">The type of the response.</typeparam>
-    public abstract class BaseEntityCreateCommandHandler<TRequest, TResponse, TEntity> : EntityCreateCommandHandler<TRequest, TResponse, TEntity, CId>
-        where TRequest : class, IRequest<TResponse>
+    public abstract class PatchCommandHandler<TRequest, TResponse, TEntity> : GenericPatchCommandHandler<TRequest, TResponse, TEntity, CId>
+        where TRequest : BaseRequest, IRequest<TResponse>
         where TEntity : BaseEntity
         where TResponse : BaseResponse
     {
@@ -26,7 +27,7 @@ namespace TurtlePath.Commands
         /// Initializes a new instance of this class.
         /// </summary>
         /// <param name="serviceProvider">The service provider used to resolve dependencies.</param>
-        protected BaseEntityCreateCommandHandler(IServiceProvider serviceProvider) : base(serviceProvider)
+        protected PatchCommandHandler(IServiceProvider serviceProvider) : base(serviceProvider)
         {
         }
     }