+++ src/TurtlePath/Commands/UpdateCommandHandler.cs
@@ -4,29 +4,33 @@ namespace TurtlePath.Commands
     using TurtlePath.Exceptions;
     using TurtlePath.Hooks;
     using TurtlePath.Models.Requests;
+    using TurtlePath.Models.Responses;
     using TurtlePath.Persistence;
     using TurtlePath.Validation;
     using TurtlePath.Mapping;
     using TurtlePath.Domain.Contracts;
     using TurtlePath.Domain.Identifier;
     using Pelican.Mediator;
+    using System;
+    using System.Threading;
+    using System.Threading.Tasks;
 
     /// <summary>
-    /// Provides a base implementation for handling delete commands for TurtlePath BaseEntity instances with CId identifiers.
+    /// Provides a base implementation for handling update commands for TurtlePath BaseEntity instances with CId identifiers.
     /// </summary>
     /// <typeparam name="TRequest">The type of the request.</typeparam>
-    /// <typeparam name="TEntity">The type of the entity being deleted.</typeparam>
+    /// <typeparam name="TEntity">The type of the entity being updated.</typeparam>
     /// <typeparam name="TResponse">The type of the response.</typeparam>
-    public abstract class BaseEntityDeleteCommandHandler<TRequest, TResponse, TEntity> : EntityDeleteCommandHandler<TRequest, TResponse, TEntity, CId>
+    public abstract class UpdateCommandHandler<TRequest, TResponse, TEntity> : GenericUpdateCommandHandler<TRequest, TResponse, TEntity, CId>
         where TRequest : BaseRequest, IRequest<TResponse>
-        where TResponse : class
+        where TResponse : BaseResponse
         where TEntity : BaseEntity
     {
         /// <summary>
         /// Initializes a new instance of this class.
         /// </summary>
         /// <param name="serviceProvider">The service provider used to resolve dependencies.</param>
-        protected BaseEntityDeleteCommandHandler(IServiceProvider serviceProvider) : base(serviceProvider)
+        protected UpdateCommandHandler(IServiceProvider serviceProvider) : base(serviceProvider)
         {
         }
     }