+++ src/TurtlePath/Queries/GetOneQueryHandler.cs
@@ -1,6 +1,7 @@
 namespace TurtlePath.Queries
 {
     using Microsoft.Extensions.DependencyInjection;
+    using TurtlePath.Exceptions;
     using TurtlePath.Hooks;
     using TurtlePath.Models.Responses;
     using TurtlePath.Persistence;
@@ -8,36 +9,36 @@ namespace TurtlePath.Queries
     using TurtlePath.Domain.Identifier;
     using Pelican.Mediator;
     using System.Linq.Expressions;
-    using System.Threading;
-    using System.Threading.Tasks;
 
     /// <summary>
-    /// Represents a query to retrieve multiple TurtlePath BaseEntity instances.
+    /// Represents a query to retrieve a single TurtlePath BaseEntity by a specified value.
     /// </summary>
+    /// <typeparam name="TValue">The type of the value used to retrieve the entity.</typeparam>
     /// <typeparam name="TEntity">The type of the entity.</typeparam>
     /// <typeparam name="TResponse">The type of the response.</typeparam>
-    public abstract class BaseEntityGetManyQuery<TEntity, TResponse> : EntityGetManyQuery<TEntity, TResponse, CId>
+    public abstract class GetOneQuery<TValue, TEntity, TResponse> : GenericGetOneQuery<TValue, TEntity, TResponse, CId>
         where TEntity : BaseEntity
         where TResponse : BaseResponse
     {
     }
 
     /// <summary>
-    /// Provides a base implementation for handling queries that retrieve multiple TurtlePath BaseEntity instances.
+    /// Provides a base implementation for handling queries that retrieve a single TurtlePath BaseEntity by a specified value.
     /// </summary>
     /// <typeparam name="TQuery">The type of the query.</typeparam>
+    /// <typeparam name="TValue">The type of the value used to retrieve the entity.</typeparam>
     /// <typeparam name="TEntity">The type of the entity.</typeparam>
     /// <typeparam name="TResponse">The type of the response.</typeparam>
-    public abstract class BaseEntityGetManyQueryHandler<TQuery, TEntity, TResponse> : EntityGetManyQueryHandler<TQuery, TEntity, TResponse, CId>
+    public abstract class GetOneQueryHandler<TQuery, TValue, TEntity, TResponse> : GenericGetOneQueryHandler<TQuery, TValue, TEntity, TResponse, CId>
+        where TQuery : GetOneQuery<TValue, TEntity, TResponse>
         where TEntity : BaseEntity
         where TResponse : BaseResponse
-        where TQuery : BaseEntityGetManyQuery<TEntity, TResponse>
     {
         /// <summary>
         /// Initializes a new instance of this class.
         /// </summary>
         /// <param name="serviceProvider">The service provider used to resolve dependencies.</param>
-        protected BaseEntityGetManyQueryHandler(IServiceProvider serviceProvider) : base(serviceProvider)
+        protected GetOneQueryHandler(IServiceProvider serviceProvider) : base(serviceProvider)
         {
         }
     }