+++ src/TurtlePath/Queries/GetPagedInfoQueryHandler.cs
@@ -14,7 +14,7 @@ namespace TurtlePath.Queries
     /// </summary>
     /// <typeparam name="TEntity">The type of the entity.</typeparam>
     /// <typeparam name="TResponse">The type of the response.</typeparam>
-    public abstract class BaseEntityGetPagedInfoQuery<TEntity, TResponse> : EntityGetPagedInfoQuery<TEntity, TResponse, CId>
+    public abstract class GetPagedInfoQuery<TEntity, TResponse> : GenericGetPagedInfoQuery<TEntity, TResponse, CId>
         where TEntity : BaseEntity
         where TResponse : BaseResponse
     {
@@ -22,7 +22,7 @@ namespace TurtlePath.Queries
         /// Initializes a new instance of this class.
         /// </summary>
         /// <param name="pagedSettings">The paged settings for the query.</param>
-        protected BaseEntityGetPagedInfoQuery(PagedSettings pagedSettings) : base(pagedSettings)
+        protected GetPagedInfoQuery(PagedSettings pagedSettings) : base(pagedSettings)
         {
         }
     }
@@ -33,8 +33,8 @@ namespace TurtlePath.Queries
     /// <typeparam name="TQuery">The type of the query.</typeparam>
     /// <typeparam name="TEntity">The type of the entity.</typeparam>
     /// <typeparam name="TResponse">The type of the response.</typeparam>
-    public abstract class BaseEntityGetPagedInfoQueryHandler<TQuery, TEntity, TResponse> : EntityGetPagedInfoQueryHandler<TQuery, TEntity, TResponse, CId>
-        where TQuery : BaseEntityGetPagedInfoQuery<TEntity, TResponse>
+    public abstract class GetPagedInfoQueryHandler<TQuery, TEntity, TResponse> : GenericGetPagedInfoQueryHandler<TQuery, TEntity, TResponse, CId>
+        where TQuery : GetPagedInfoQuery<TEntity, TResponse>
         where TEntity : BaseEntity
         where TResponse : BaseResponse
     {
@@ -42,7 +42,7 @@ namespace TurtlePath.Queries
         /// Initializes a new instance of this class.
         /// </summary>
         /// <param name="serviceProvider">The service provider used to resolve dependencies.</param>
-        protected BaseEntityGetPagedInfoQueryHandler(IServiceProvider serviceProvider) : base(serviceProvider)
+        protected GetPagedInfoQueryHandler(IServiceProvider serviceProvider) : base(serviceProvider)
         {
         }
     }